Create A rational choice
Input: As input, the program receives the number of advertisements n. The following lines contain n descriptions of advertisements. Each description consists of the name of the car in the first line of the description and two numbers p and d in the second line, where p is a floating-point number and denotes the car's mileage in km, and d is an integer and denotes the wheel diameter in inches.
Output: At the output, the program should print 3 names of cars with the smallest number of complete wheel revolutions along with this number. The data should be entered in ascending order of the number of revolutions. For an equal number of revolutions, first enter the data of the car whose name is earlier in alphabetical order.