Create "Święto kwitnącej fasoli"
Input: As standard input, the program will receive song lyrics containing letters of the Latin alphabet. The end of the text will be marked with the ^ symbol. The length of the song lyrics will not exceed 2,000,000 characters.
Output: On the standard output, the program should print the most common and least frequent letters separated by a space. When searching for the rarest letter, the program should only take into account letters that appear at least once in the song. If several letters appear the same number of times, write the letter that is earlier in the alphabet. The program should not distinguish between uppercase and lowercase letters when counting. The result should be lowercase.