Create Recruitment
Input: The first line of input contains a natural number t, i.e. the number of data sets. Each data set consists of two natural numbers n - the number of names and p - the number of names that should be displayed in the output. The following lines contain descriptions of people. The description of a person consists of: the name as a one-word string of characters and the natural number x, denoting the number of minutes required to convince this person to cooperate.
Output: The output of the program for each data set is to print the first p people that the Partition should visit and the minimum time in minutes that should be spent on recruiting helpers. The result should be the remainder when divided by 10^6+3. (No stl libraries)