Create SortOfPairs
Input : The input is an integer n. The second line of the input contains three letters separated by spaces: 's', 'd' and 'm', and their result is to determine the order of the children. In rows n lines there are poems describing subsequent students of schools. the line of lines consists of the description of two boys: girls and girls. The student's description consists of a one-word string encoding the student's code, one floating-point number s denoting the definition, two integer units d and m denoting multiple distances from the place of residence in meters, and the birth number of the student. Months are numbered from 1.
Output: At the output, the program should print ordered lists of girls and boys in pairs according to the given order of criteria. The pair should be displayed in the form: code1 code2. The entire list should be displayed in one line. For two students whose corresponding values ​​are all equal, maintain the order of occurrence in the list.