Create Cast.cpp
Input: The input data consists of a natural number 1<=n<=100000, denoting the number of operations, and in the following lines the number of the operation op={1,2}, performed on the available actor database, along with a description.
*) 1 - adding an actor. The description of the operation consists of 2 lines, where the first one is the name(s) and surname of the actor or his nickname, and the second is a natural number f, where f is the number of the actor's fans who will stop watching the series after his possible death.
*) 2 - finding an actor for the role. The description of the operation consists of a natural number p, denoting the maximum number of fans that you can afford to lose in the event of the death of this character.
Output: At the output, the program should print the result of the mentioned operations, where:
*) for the operation of adding an actor (1), the string should be "YES" if the actor was successfully added and the string "NO" if the same actor with the same number of fans already exists in the database.
*) for the operation of finding an actor (2) it will be the personal data of a selected actor with the largest possible number of fans, whose death in turn will not exceed the allowable losses, or the inscription "NO" if no actor meets the criterion.

It is obvious that once an actor receives a role, he cannot get another role in the series. If more than one actor meets the criterion, the actor with the lexicographically smallest name and surname should be listed.